refactor(apple): relocate RemoteControlCoordinator persistence helpers (pure move, LOC headroom) (#1423)
persistPaired(token:resolved:) and touchLastConnected() move
byte-identically into a new RemoteControlCoordinator+Persistence.swift,
mirroring the +UIPhase.swift precedent (#1424) — this frees the LOC
headroom the pocket-control redesign's six upcoming relay-tap/retire-hook
call-lines need to land in RemoteControlCoordinator.swift without
breaching the 400-line budget.

store/rebuildRows() promote from private to internal access on
RemoteControlCoordinator so the sibling file can reach them, the same
cross-file convention already used for session/currentTVName/
savedRecords. RemoteControlUIStateTests and the rest of the suite are
unchanged and stay green.
